add integration test & reorganize
