+++ api/tests/opentrons/protocol_api_integration/test_liquid_classes.py
@@ -1,5 +1,5 @@
 """Tests for the APIs around liquid classes."""
-from typing import Dict
+from typing import Dict, Any
 
 import pytest
 
@@ -69,6 +69,39 @@ def test_custom_liquid_class_creation_and_property_fetching(
     custom_water_props = custom_water.get_for(p50, tiprack)
     assert custom_water_props.aspirate.submerge.speed == 100
     assert custom_water_props.dispense.flow_rate_by_volume.get_for_volume(20) == 30
+
+    with pytest.raises(
+        ValueError,
+        match="No properties found for flex_8channel_50 in water_50 liquid class",
+    ):
+        custom_water.get_for(
+            "flex_8channel_50", "opentrons/opentrons_flex_96_tiprack_50ul/1"
+        )
+
+
+@pytest.mark.ot3_only
+@pytest.mark.parametrize(
+    "simulated_protocol_context", [("2.24", "Flex")], indirect=True
+)
+def test_custom_liquid_class_w_multi_dispense_creation_and_property_fetching(
+    simulated_protocol_context: ProtocolContext,
+    custom_pip_n_tip_transfer_properties_dict_v2: Dict[str, Dict[str, Any]],
+) -> None:
+    """It should create the liquid class with specified properties and provide access to its properties."""
+    p50 = "a_custom_pipette_type"
+    tiprack = "a_custom_tiprack_uri"
+    custom_water = simulated_protocol_context.define_liquid_class(
+        name="water_50",
+        properties=custom_pip_n_tip_transfer_properties_dict_v2,
+        display_name="Custom Aqueous",
+    )
+
+    assert custom_water.name == "water_50"
+    assert custom_water.display_name == "Custom Aqueous"
+    custom_water_props = custom_water.get_for(p50, tiprack)
+    assert custom_water_props.aspirate.retract.speed == 40
+    assert custom_water_props.dispense.submerge.speed == 100
+    assert custom_water_props.multi_dispense.retract.blowout.location.value == "trash"  # type: ignore[union-attr]
 
     with pytest.raises(
         ValueError,
